+++ NEWS.md
@@ -20,6 +20,8 @@ for human readability.
 - New structured, curvilinear, conforming mesh type `CurvedMesh` (experimental)
 - New unstructured, curvilinear, conforming mesh type `UnstructuredQuadMesh` in 2D (experimental)
 - New unstructured, curvilinear, adaptive (non-conforming) mesh type `P4estMesh` in 2D (experimental)
+- Experimental support for finite difference (FD) summation-by-parts (SBP) methods via
+  [SummationByPartsOperators.jl](https://github.com/ranocha/SummationByPartsOperators.jl)
 
 #### Changed